pyproject.toml
[project]
name = "Image_Hosting"
version = "0.1.0"
description = "Image Hosting for Private Server"
authors = [
    {name = "Alexandr Altukhov"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "legacy-cgi (>=2.6.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
